chore: Add FlexNode Update Hook
