+++ src/components/shared/ReactFlow/FlowCanvas/FlexNode/FlexNodeEditor.tsx
@@ -11,15 +11,13 @@ import { BlockStack, InlineStack } from "@/components/ui/layout";
 import { Textarea } from "@/components/ui/textarea";
 import { Paragraph, Text } from "@/components/ui/typography";
 import { cn } from "@/lib/utils";
-import { useComponentSpec } from "@/providers/ComponentSpecProvider";
 import { FONT_SIZE_MD, FONT_SIZE_SM } from "@/utils/constants";
-import { updateSubgraphSpec } from "@/utils/subgraphUtils";
 
 import { StackingControls } from "../../FlowControls/StackingControls";
-import { updateFlexNodeInComponentSpec } from "./interface";
 import LockToggle from "./LockToggle";
 import { TextSizeSelector } from "./TextSizeSelector";
 import type { FlexNodeData } from "./types";
+import { useFlexNodeUpdate } from "./useFlexNodeUpdate";
 import { DEFAULT_BORDER_COLOR } from "./utils";
 
 interface FlexNodeEditorProps {
@@ -108,12 +106,7 @@ const ContentEditor = ({
   flexNode: FlexNodeData;
   readOnly: boolean;
 }) => {
-  const {
-    componentSpec,
-    currentSubgraphSpec,
-    currentSubgraphPath,
-    setComponentSpec,
-  } = useComponentSpec();
+  const { updateProperties } = useFlexNodeUpdate(flexNode);
 
   const { properties } = flexNode;
 
@@ -129,67 +122,15 @@ const ContentEditor = ({
   };
 
   const handleTitleFontSizeChange = (newSize: number) => {
-    const updatedSubgraphSpec = updateFlexNodeInComponentSpec(
-      currentSubgraphSpec,
-      {
-        ...flexNode,
-        properties: {
-          ...properties,
-          titleFontSize: newSize,
-        },
-      },
-    );
-
-    const newRootSpec = updateSubgraphSpec(
-      componentSpec,
-      currentSubgraphPath,
-      updatedSubgraphSpec,
-    );
-
-    setComponentSpec(newRootSpec);
+    updateProperties({ titleFontSize: newSize });
   };
 
   const handleContentFontSizeChange = (newSize: number) => {
-    const updatedSubgraphSpec = updateFlexNodeInComponentSpec(
-      currentSubgraphSpec,
-      {
-        ...flexNode,
-        properties: {
-          ...properties,
-          contentFontSize: newSize,
-        },
-      },
-    );
-
-    const newRootSpec = updateSubgraphSpec(
-      componentSpec,
-      currentSubgraphPath,
-      updatedSubgraphSpec,
-    );
-
-    setComponentSpec(newRootSpec);
+    updateProperties({ contentFontSize: newSize });
   };
 
   const saveChanges = () => {
-    const updatedSubgraphSpec = updateFlexNodeInComponentSpec(
-      currentSubgraphSpec,
-      {
-        ...flexNode,
-        properties: {
-          ...properties,
-          title,
-          content,
-        },
-      },
-    );
-
-    const newRootSpec = updateSubgraphSpec(
-      componentSpec,
-      currentSubgraphPath,
-      updatedSubgraphSpec,
-    );
-
-    setComponentSpec(newRootSpec);
+    updateProperties({ title, content });
   };
 
   useEffect(() => {
@@ -288,12 +229,7 @@ const ColorEditor = ({
   flexNode: FlexNodeData;
   readOnly: boolean;
 }) => {
-  const {
-    componentSpec,
-    currentSubgraphSpec,
-    currentSubgraphPath,
-    setComponentSpec,
-  } = useComponentSpec();
+  const { updateProperties } = useFlexNodeUpdate(flexNode);
 
   const { properties } = flexNode;
 
@@ -320,25 +256,10 @@ const ColorEditor = ({
     newBackgroundColor: string,
     newBorderColor: string | undefined,
   ) => {
-    const updatedSubgraphSpec = updateFlexNodeInComponentSpec(
-      currentSubgraphSpec,
-      {
-        ...flexNode,
-        properties: {
-          ...properties,
-          color: newBackgroundColor,
-          borderColor: newBorderColor,
-        },
-      },
-    );
-
-    const newRootSpec = updateSubgraphSpec(
-      componentSpec,
-      currentSubgraphPath,
-      updatedSubgraphSpec,
-    );
-
-    setComponentSpec(newRootSpec);
+    updateProperties({
+      color: newBackgroundColor,
+      borderColor: newBorderColor,
+    });
   };
 
   useEffect(() => {
@@ -409,29 +330,10 @@ const ColorEditor = ({
 };
 
 const ZIndexEditor = ({ flexNode }: { flexNode: FlexNodeData }) => {
-  const {
-    componentSpec,
-    currentSubgraphSpec,
-    currentSubgraphPath,
-    setComponentSpec,
-  } = useComponentSpec();
+  const { updateFlexNode } = useFlexNodeUpdate(flexNode);
 
   const handleStackingControlChange = (newZIndex: number) => {
-    const updatedSubgraphSpec = updateFlexNodeInComponentSpec(
-      currentSubgraphSpec,
-      {
-        ...flexNode,
-        zIndex: newZIndex,
-      },
-    );
-
-    const newRootSpec = updateSubgraphSpec(
-      componentSpec,
-      currentSubgraphPath,
-      updatedSubgraphSpec,
-    );
-
-    setComponentSpec(newRootSpec);
+    updateFlexNode({ zIndex: newZIndex });
   };
 
   return (
